moves UserChangeLogger under users
